Feat : Simple validation OwO

+++ app/Http/Controllers/ShortLinkController.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace App\Http\Controllers;
 
 use Illuminate\Support\Str;
@@ -8,6 +7,20 @@ use App\Models\ShortLink;
 
 class ShortLinkController extends Controller
 {
+    protected $protectedCodes = [
+        'masfana',
+        'lawos',
+        'repoini',
+    ];
+
+    protected function Jaga(string $code)
+    {
+        if (in_array($code, $this->protectedCodes)) {
+            return redirect()->back()->with('error', 'Mangsud amat le :3');
+        }
+        return false;
+    }
+
     public function index()
     {
         $shortLinks = ShortLink::latest()->get();
@@ -16,6 +29,7 @@ class ShortLinkController extends Controller
 
     public function store(Request $request)
     {
+
         $request->validate([
             'code' => 'nullable|unique:short_links,code',
             'url' => 'required|url|unique:short_links,url|max:2048'
@@ -23,7 +37,7 @@ class ShortLinkController extends Controller
 
         ShortLink::create([
             'url' => $request->url,
-            'code' =>  $request->code ?? Str::random(6),
+            'code' => $request->code ?? Str::random(6),
             'click_count' => 0
         ]);
 
@@ -32,11 +46,23 @@ class ShortLinkController extends Controller
 
     public function edit(ShortLink $shortLink)
     {
+        $check = $this->Jaga($shortLink->code);
+        if ($check) {
+            return $check;
+        }
+
         return view('edit', compact('shortLink'));
+
     }
 
     public function update(Request $request, ShortLink $shortLink)
     {
+
+        $check = $this->Jaga($shortLink->code);
+        if ($check) {
+            return $check;
+        }
+
         $request->validate([
             'code' => 'required|unique:short_links,code',
             'url' => 'required|url|max:2048'
@@ -46,12 +72,17 @@ class ShortLinkController extends Controller
             'url' => $request->url,
             'code' => $request->code
         ]);
-        
+
         return redirect()->route('shorten.link.index')->with('success', 'Short link updated successfully!');
     }
 
     public function destroy(ShortLink $shortLink)
     {
+        $check = $this->Jaga($shortLink->code);
+        if ($check) {
+            return $check;
+        }
+
         $shortLink->delete();
         return redirect()->back()->with('success', 'Short link deleted successfully!');
     }
